feat: endpointtests

+++ tests/functional/src/test_roles.py
@@ -1,4 +1,3 @@
-
 from http import HTTPStatus
 
 
@@ -6,11 +5,13 @@ def test_create_role(client_with_db, access_headers, create_user):
     """
     Тестирование создания роли
     """
-    new_role = 'test_create_role'
-    response = client_with_db.post('/v1/create_role',
-                                   data={"new_role": new_role},
-                                   headers=access_headers)
+    new_role = "test_create_role"
+    response = client_with_db.post("/v1/create_role", data={"new_role": new_role}, headers=access_headers)
     assert response.status_code == HTTPStatus.OK
+    response = client_with_db.post("/v1/create_role", data={"new_role": new_role}, headers=access_headers)
+    assert response.status_code == HTTPStatus.INTERNAL_SERVER_ERROR
+    response = client_with_db.post("/v1/create_role", data={"new_role": ""}, headers=access_headers)
+    assert response.status_code == HTTPStatus.UNAUTHORIZED
 
 
 def test_delete_role(client_with_db, create_role, access_headers):
@@ -18,29 +19,32 @@ def test_delete_role(client_with_db, create_role, access_headers):
     Тестирование удаления роли
     """
     role = create_role.name
-    response = client_with_db.delete('/v1/delete_role',
-                                     data={"role": role},
-                                     headers=access_headers)
+    response = client_with_db.delete("/v1/delete_role", data={"role": role}, headers=access_headers)
     assert response.status_code == HTTPStatus.OK
+    response = client_with_db.delete("/v1/delete_role", data={"role": role}, headers=access_headers)
+    assert response.status_code == HTTPStatus.CONFLICT
 
 
-def test_change_role(client_with_db, create_role, access_headers):
+def test_rename_role(client_with_db, create_role, access_headers):
     """
     Тестирование изменение роли
     """
     role = create_role.name
-    new_name = 'test_change_role'
-    response = client_with_db.put('/v1/change_role',
-                                  data={"role": role,
-                                        'new_name': new_name},
-                                  headers=access_headers)
+    new_name = "test_rename_role"
+    response = client_with_db.put("/v1/rename_role", data={"role": role, "new_name": new_name}, headers=access_headers)
     assert response.status_code == HTTPStatus.OK
+    not_valid_role = "not_role"
+    response = client_with_db.put(
+        "/v1/rename_role", data={"role": not_valid_role, "new_name": new_name}, headers=access_headers
+    )
+    assert response.status_code == HTTPStatus.CONFLICT
 
 
-def test_roles_list(client_with_db, create_role, access_headers):
+def test_roles_list(client_with_db, create_role, access_headers, not_access_headers):
     """
     Тестирование вывода списка ролей
     """
-    response = client_with_db.get('/v1/roles_list',
-                                  headers=access_headers)
+    response = client_with_db.get("/v1/roles_list", headers=access_headers)
     assert response.status_code == HTTPStatus.OK
+    response = client_with_db.get("/v1/roles_list", headers=not_access_headers)
+    assert response.status_code == HTTPStatus.UNAUTHORIZED
